Finalized framework-api module for positive tests
Added all the code required for supporting valid and working queries execution. Currently, the module does not handle the cases when there are errors during the API calls properly, but this is something that will be added in the future commits.

General approach to working with remote resources (aka user, post ...) is to create an instance of Request<HttpMethod> and to pass it down to the request executor. The executor will return Response<Resource> which not only contains HTTP details, but also the Resource object itself.